chore: Refactor Navbar component and related files
