+++ src/components/Navbar/NavbarToggle.jsx
@@ -1,3 +1,4 @@
+import { Toggle, SvgX } from "../../utils";
 export const NavbarToggle = ({ onClick, isOpen }) => (
   <div className="flex items-center  justify-end ">
     <div className="sm:hidden">
@@ -9,40 +10,7 @@ export const NavbarToggle = ({ onClick, isOpen }) => (
         aria-controls="navbar-collapse-with-animation"
         aria-label="Toggle navigation"
       >
-        {isOpen ? (
-          <svg
-            className="flex-shrink-0 size-4"
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          >
-            <path d="M18 6 6 18" />
-            <path d="m6 6 12 12" />
-          </svg>
-        ) : (
-          <svg
-            className="flex-shrink-0 size-4"
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          >
-            <line x1="3" x2="21" y1="6" y2="6" />
-            <line x1="3" x2="21" y1="12" y2="12" />
-            <line x1="3" x2="21" y1="18" y2="18" />
-          </svg>
-        )}
+        {isOpen ? <Toggle /> : <SvgX />}
       </button>
     </div>
   </div>
