+++ src/hooks/index.jsx
@@ -5,3 +5,4 @@ export { useEmail } from "./useEmail";
 export { useProject } from "./useProject";
 export { useFilterProjects } from "./useFilterProjects";
 export { useLoading } from "./useLoading";
+export { useNavbar } from "./useNavbar";
